fix clearing dicom

+++ api/pacs.py
@@ -14,6 +14,7 @@ import pandas as pd
 import datetime
 from zipfile import ZipFile
 from Constant import MONGO_URL, SECRET
+from pathlib import Path
 
 import os, sys
 import json
@@ -164,58 +165,55 @@ async def clear_dicom_folder():
     LOG_DIR = os.path.join(BASE_DIR, "resources", "log", "log_receive_dcm")
     today_date = datetime.date.today()
     try:
-        delete_date = today_date
-        n = today_date.day
-        if today_date.day == 28: # delete dicom received before 14th
-            delete_date = today_date.replace(day=14)
-            n = 28
-        elif today_date.day == 14: # delete dicom received before 28th
-            delete_date = today_date.replace(day=28, month=today_date.month-1)
-            n = 14
-        while delete_date.day != n:
-            year = delete_date.year
-            month = delete_date.month
+        print(str(today_date), 'Start clearing dicom process')
+        # delete dicom (2 weeks)
+        delete_date = today_date - datetime.timedelta(days=15) 
+        year = str(delete_date.year)
+        month = str(delete_date.month)
+
+        folder_path = os.path.join(BASE_DIR, 'resources', 'log', 'log_clear_dicom', 'delete_dicom', year, month)
+        Path(folder_path).mkdir(parents=True, exist_ok=True)
+
+        log_delete_dicom_path = os.path.join(folder_path, str(today_date)+'.csv')
+        if os.path.exists(os.path.join(LOG_DIR, year, month, str(delete_date) + '.csv')):
+            # print(str(delete_date) + '.csv')
             df = pd.read_csv(os.path.join(LOG_DIR, year, month, str(delete_date) + '.csv'))
             unique_acc_no = set(df["Accession Number"])
             # check acc no in mongo
             db = pymongo.MongoClient(MONGO_URL)["webapp"]
             used_acc_no = db["images"].distinct("accession_no", {"accession_no": {"$in": list(unique_acc_no)}})
+            # print(used_acc_no, unique_acc_no)
             for acc_no in unique_acc_no:
+                log_data = {'Accession Number': acc_no, "Success": ""}
                 if acc_no not in used_acc_no and os.path.exists(os.path.join(PACS_DIR, acc_no + '.evt')):
                     os.remove(os.path.join(PACS_DIR, acc_no + '.evt'))
-            delete_date = delete_date - datetime.timedelta(days=1)
+                    log_data = {'Accession Number': acc_no, "Success": True}
+                    # print(log_data)
+                    log_data = pd.DataFrame.from_records([log_data])
+                    if not os.path.exists(log_delete_dicom_path):
+                        log_data.to_csv(log_delete_dicom_path, index=False)
+                    else:
+                        log_data.to_csv(log_delete_dicom_path, index=False, mode='a', header=False)
         
-        if today_date.day == 28: # change original dicom to dummy dicom
-            delete_date = today_date.replace(day=28, month=today_date.month-1)
-            while True:
-                year = delete_date.year
-                month = delete_date.month
-                df = pd.read_csv(os.path.join(LOG_DIR, year, month, str(delete_date) + '.csv'))
-                unique_acc_no = set(df["Accession Number"])
-                used_acc_no = db["images"].distinct("accession_no", {
-                    "accession_no": {"$in": list(unique_acc_no)},
-                    "hn": {"$ne": ""}
-                    })
-                subprocess.run(["python", os.path.join(BASE_DIR, "pacs_connection", "dicom_function.py"), "-f", "get_dummy", "-l", used_acc_no])
-                delete_date = delete_date - datetime.timedelta(days=1)
-
-                image_id_list = db["images"].find({"accession_no": {"$in": used_acc_no}}, {"_id"})
-                result_list = db["pred_results"].find({"image_id", {"$in": image_id_list}})
-                for result in result_list:
-                    db["pred_results"].find_one_and_update({"_id": result["_id"]}, {
-                        "hn": None,
-                        "patient_name": None
-                    })
-                    db["images"].find_one_and_update({"_id": result["image_id"]}, {
-                        "hn": None,
-                    })
-                    db["medrecords"].find_one_and_update({"_id": result["record_id"]}, {
-                        "record.hn": None,
-                    })
-                if delete_date.day == 28:
-                    break
-        del df
-        return JSONResponse(content={"success": True, "message": "Finish clearing dicom"}, status_code=200)
+        # change original dicom to dummy dicom (1 month)
+        dummy_date = today_date - datetime.timedelta(days=31)  
+        year = str(dummy_date.year)
+        month = str(dummy_date.month)
+        used_acc_no = []
+        # print(dummy_date)
+        if os.path.exists(os.path.join(LOG_DIR, year, month, str(dummy_date) + '.csv')):
+            print(str(dummy_date) + '.csv')
+            df = pd.read_csv(os.path.join(LOG_DIR, year, month, str(dummy_date) + '.csv'))
+            unique_acc_no = set(df["Accession Number"])
+            used_acc_no = db["images"].distinct("accession_no", {
+                "accession_no": {"$in": list(unique_acc_no)},
+                "hn": {"$ne": None}
+                })
+            # print(unique_acc_no, used_acc_no)
+            if used_acc_no != []:
+                subprocess.run(["python", os.path.join(BASE_DIR, "pacs_connection", "dicom_function.py"), "-f", "convert_evt_to_dummy", "-l", " ".join(used_acc_no)])
+        
+        return JSONResponse(content={"success": True, "message": "Finish clearing dicom", "data": used_acc_no}, status_code=200)
     except Exception as e:
         print(traceback.format_exc())
         return JSONResponse(content={"success": False, "message": e}, status_code=500)
